+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+## [0.1.0-beta.4]
+
 ### Added
 
 - feat: Use fastly command from `@fastly/cli` package
@@ -46,7 +48,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
-[unreleased]: https://github.com/fastly/create-compute-js/compare/v0.1.0-beta.3...HEAD
+[unreleased]: https://github.com/fastly/create-compute-js/compare/v0.1.0-beta.4...HEAD
+[0.1.0-beta.4]: https://github.com/fastly/create-compute-js/compare/v0.1.0-beta.3...v0.1.0-beta.4
 [0.1.0-beta.3]: https://github.com/fastly/create-compute-js/compare/v0.1.0-beta.2...v0.1.0-beta.3
 [0.1.0-beta.2]: https://github.com/fastly/create-compute-js/compare/v0.1.0-beta.1...v0.1.0-beta.2
 [0.1.0-beta.1]: https://github.com/fastly/create-compute-js/compare/v0.1.0-beta.0...v0.1.0-beta.1